api: move itunnel handlers to separate file
